Move communication component to components folder
