+++ src/components/ChatBox.tsx
@@ -1,0 +1,58 @@
+import useSessionComm from "~/hooks/useSessionComm";
+
+const Chatbox = ({
+  sessionId,
+  userId,
+  userName,
+  className,
+}: {
+  sessionId: string;
+  userId: string;
+  userName: string;
+  className?: string | undefined;
+}) => {
+  const [
+    allSessionMessages,
+    sendMessage,
+    onTyping,
+    currentMessage,
+    userTypingMessage,
+  ] = useSessionComm(sessionId, userId, userName);
+
+  return (
+    <div className={className}>
+      <div className="messages-container overflow-y-auto flex flex-col h-[18.5rem]">
+        {allSessionMessages?.map((message, index) => {
+          return (
+            <div
+              key={index}
+              className="w-full rounded-md dark:bg-gray-700 text-white p-2 my-2"
+            >
+              <div className="flex justify-between">
+                <span>{message.senderName}</span>
+                <span>Sent at {message.createdAt.toLocaleTimeString()}</span>
+              </div>
+              <p className="w-[27.875rem] whitespace-normal break-words">
+                {message.message}
+              </p>
+            </div>
+          );
+        })}
+      </div>
+      <div className="flex flex-col mt-2">
+        <p className="text-white mb-2">{userTypingMessage}</p>
+        <div className="flex">
+          <input
+            className="w-full rounded-md p-2 focus:ring-primary-600 focus:border-primary-600"
+            type="text"
+            value={currentMessage}
+            onChange={(e) => onTyping(e.target.value)}
+            onKeyDown={(e) => e.key === "Enter" && sendMessage()}
+          />
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Chatbox;
